Added sky, floor, and preliminary algorithm for drawing the 3D effect. Also changed the command structure and added keyboard input.

+++ src/com/joshuacrotts/raycaster/main/Wall.java
@@ -36,7 +36,7 @@ public class Wall {
       return null;
     }
 
-    int denominator = (int)((y4 - y3) * (x2 - x1) - (x4 - x3) * (y2 - y1));
+    double denominator = ((y4 - y3) * (x2 - x1) - (x4 - x3) * (y2 - y1));
 
     // Lines are parallel
     if (denominator == 0) {
